feat: reorganizing folders order
- Puts week_1 folder in main folder

style: changes file name from all_classes to first_steps

style: moves src/week_2/class_1 to src/main/week_2/class_1

style: moves week_2/class_2/stack tests to test/week_2/class_2/stack folder

style: moves queue src and test to ./src/ path
